backend/requirements.txt
Django==4.1
djangorestframework==3.13
psycopg2-binary
django-money
